requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
MarkupSafe==2.0.1
flask-restful
passlib
email-validator
mysql-connector-python
flask-jwt-extended
requests
